requirements.txt
wordcloud==1.8.1
PyKomoran==0.1.5
python-mecab-ko==1.0.11


